requirements.txt
langchain
unstructured
tiktoken
pinecone-client
pypdf
openai
pandas
numpy
python-dotenv
pdfreader
langchain-community
cassandra
